mptcp: add FastOpen sender tests
For the moment, only the client part is validated.

2 methods are currently tested:
- TCP_FASTOPEN_CONNECT
- MSG_FASTOPEN

For each method, there are two tests:
- one without using TFO cookies
- one where a first connection is created to validate the cookie part.
  Then a second connection is established with data in the SYN.

Note that it looks like Packetdrill doesn't support creating a 2nd
connection on the same script yet. When we do that, Packetdrill reports
the 3rd ACK of the 2nd connection doesn't contain the expected the
client key. But that's alright for the moment, tests have been split:
with and without cookies.

These tests can be used to validate Benjamin' series: mptcp: add support
for TFO, sender side only. In the v2, MSG_FASTOPEN support has been
removed as it is part of Dmytro's series.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>