mptcp: add_addr: more tolerant with retrans timeout after errors
ADD_ADDR restransmissions are expected to come ~1s after the last sent
packet.

In packetdrill, the timing is compared to the last packet. We would like
to tell Packetdrill that the packet is expected to come "1s not after
the previous packet but the one before" but that's not possible.

In this test the last packet before the ADD_ADDR retransmission is not
the ADD_ADDR that is sent before but it is the injected and faulty echo
ADD_ADR. With a (very) slow host, it can takes a bit of time to inject
the packet. Because of that, the retransmitted packet can arrive earlier
than expected: if the faulty echo ADD_ADDR injection takes X sec, the
retransmission of the ADD_ADDR by the kernel will take (1 - X) sec. If X
is bigger than the tolerance, the test fails which seems to happen
regularly on the public CI.

  add_addr_retry_v4.pkt:22: error handling packet: timing error: expected outbound packet at 3.214190 sec but happened at 2.384757 sec; tolerance 0.800000 sec
  script packet:  3.214190 . 1:1(0) ack 1 <add_address address_id: 1 ipv4: 192.168.0.3 hmac: 10354569113664661296>
  actual packet:  2.384757 . 1:1(0) ack 1 win 256 <add_address address_id: 1 ipv4: 192.168.0.3 hmac: 10354569113664661296>

We should then accept packets being sent less than one second before the
injected and faulty echo ADD_ADDR.

Even worst with a debug kernel config:

  add_addr_retry_v4.pkt:22: error handling packet: timing error: expected outbound packet at 10.335930 sec but happened at 5.908318 sec; tolerance 2.000000 sec
  script packet: 10.335930 . 1:1(0) ack 1 <add_address address_id: 1 ipv4: 192.168.0.3 hmac: 14923819917020444620>
  actual packet:  5.908318 . 1:1(0) ack 1 win 256 <add_address address_id: 1 ipv4: 192.168.0.3 hmac: 14923819917020444620>

Injecting the faulty echo and processing packets sent by the kernel can
take time and we have situations where packets arrive a few seconds
before the expected time by Packetdrill!

Sadly, we cannot tell Packetdrill the packet is expected to be sent in
the past. So we need to increment the tolerance a bit. But that's find
to do that because a new test has been added in the parent commit: it is
focussing on the ADD_ADDR retransmissions without injecting other
packets in between. This other test can have stricter expected time.

Closes: https://github.com/multipath-tcp/mptcp_net-next/issues/312
Suggested-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>