mptcp: tfo: add basic server side coverage
These tests are similar to the ones for the client side: a first
connection is established without DATA in the SYN to request a key, then
new connections can send DATA in the SYN with the cookie it previously
received.

Ideally we would like to validate a second connection in the same test
but we are limited by Packetdrill here which cannot understand the new
SYN is for a new connection.

So this test is split in two parts:

- server-TCP_FASTOPEN-cookie-req: the first connection, cookie request
- server-TCP_FASTOPEN-cookie-data: the cookie has already been
  exchanged, data can be sent with the cookie.

Note that we can also not use the cookie by setting a sysctl and
directly send data, that's what "server-tfo-no-cookie" is doing.

Equivalent tests for (plain) TCP are also added.

Note that kernel patches are required to get the expected cookie in
IPv6, see:

  https://github.com/multipath-tcp/mptcp_net-next/issues/317

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>