mptcp: new test to validate IPV6_UNICAST_HOPS sockopt
This sockopt is going to be supported with MPTCP [1], it is good to
validate it with packetdrill.

The value is changed, retrieved and validated by looking at the
generated packets.

The value is also modified later when the connexion is on-going.

The test is similar to the one validating IP_TTL socket option but here
for IPv6 only.

Link: https://patchwork.kernel.org/project/mptcp/patch/20230707-mptcp-unify-sockopt-issue-353-v1-6-693e15c06646@tessares.net/ [1]
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>