packetdrill: run_all: limit tests running in parallel
When using machines with limited resources (e.g. public CI) and a kernel
with a debug config, some errors have been observed when too many tests
were executed in parallel, e.g. packets being received too late, etc.

This commit allows to limit the number of tests being launched in
parallel while still keeping a timeout per test and all the other
features. If the new '-P' option is not used, the behaviour is not
modified.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>