packetdrill: run_all: add --dry_run mode
It is useful to have an easy way to display the commands that are going
to be used: for debugging purposes, to relaunch one specific test with
different arguments or from a different environment, etc.

Now a new '--dry_run' parameter is available to only display commands
and not execute anything.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>