mptcp: add missing ACK on some packets
Since the kernel commit 3d501dd326fb ("tcp: do not accept ACK of bytes
we never sent"), two MPTCP fastclose tests were no longer passing
because the wrong ACK number was being sent. As a result, the ACK packet
was dropped and the behaviour was not the expected one.

Simply setting the correct ACK number fixes the issue.

Signed-off-by: Matthieu Baerts <matttbe@kernel.org>