mptcp: add notsent_lowat support
Based on epoll_out_edge_notsent_lowat.pkt, with some adaptation to cope
with MPTCP specificities.

Note that this will cause some random failures around line 43.

The root cause is that the packetdrill core tries to aggregate eligible
TCP packets and can replace the wire packet with the aggregate one to
fit the drill under test.

The problem is that the aggregation support is partial/buggy: packets
with the same tcp options len, but different tcp options contents will
be considered eligible for aggregation, and the aggregate packets will
carry the options for the 1st pkt on the wire. See verify_tcp() in
run_packet.c.

The above works e.g. with tcp stamp options, but does not work with
mptcp: packets with different DSS are aggregated carrying the DSS from
the 1st one. Note that this is different from what the GRO engine is
doing and attempts to mimic what the TCP stack coalescing is doing. With
the major difference that such coalescing merges correctly different
DSS.

TL;DR: to solve the failures for good non trivial changes to the core
are required, but sharing the drill early to somewhat covers the pending
kernel patches.

Link: https://github.com/multipath-tcp/mptcp_net-next/issues/476
Signed-off-by: Paolo Abeni <pabeni@redat.com>
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>