packetdrill: run_all: add --tap <DIR> option
This new option will generate simple TAP files in the given directory,
for each path managed by packetdrill: the given one, except if -s is
passed.

This is particularly useful because the TAP format is used by other test
frameworks linked to the kernel: kselftests, kunit, etc. So a CI can
easily parse these files to display results in a board, and track
regressions per test.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>