bpf: add scheduler tests
This patch adds tests for the round-robin scheduler. The current
implementation sends batches of data and works well for small writes,
i.e. < MSS. But for large writes of data, the round-robin scheduler
does not, yet, sends each piece of data in alternance over each subflow.

the load_struct_ops.sh takes one parameter the bpf program to load. It
will look for this file in the kernel build directory.

Signed-off-by: Gregory Detal <gregory.detal@gmail.com>